// apps/messenger/src/models/profile.model.ts
export interface threadProfileData {
  _id: string;
  cell_phone: string;
  cellphone_verify: boolean;
  date_last_login: {
    $date: number;
  };
  email_comm: boolean;
  first_name: string;
  image_id: string;
  last_name: string;
  profile: UserProfile;
  sms_comm: boolean;
  status: status;
  type: string;   
  flag_active?:boolean;
}

export interface UserProfile {
  dept: string;
  iid: string;
  iname: string;
  ishort: string;
  pager_number: string;
  title: string;
}
export interface status {
  away_message?: string;
  away_message_mode: string;
  is_signed_out: boolean;
  r?: [];
  r_type?: string[];
  s?: string;
  c?: {
    ref: string;
    type: string;
  };
  role_notify?:ROLE_NOTIFY[];
}
// ...
export interface USER_PROFILE_THREAD {
  threads: thread;
  profiles: threadProfileData[];
  sender_profile: threadProfileData | undefined
}
export interface thread {
patient: any;
patient_name: any;
  _id: string;
  iid: string;
  origin: string;
  seq: number;
  expiration?: string;
  time_created: string;
  time_updated: string;
  type: string;
  unread_count: number;
  subject: string;
  recipients:string[] ;
  messages: Message[];
  user_id: string;
  status?:string;
  receivers: Array<{ name: string }>;
  visibility?: boolean;
  muted?: boolean;
  isOlderThread?: boolean;

}
// ...
export interface Message {
  _id: {
    $oid: string;
  };
  content: string;
  iid: string;
  seq: number;
  status: string;
  statuses: MessageStatus[];
  thread_id: {
    $oid: string;
  };
  time_created: {
    $date: number;
  };
  type: string;
  user_id: string;
  urgent?: boolean;

}
export interface MessageStatus {
  _id: {
    $oid: string;
  };
  message_id: {
    $oid: string;
  };
  status: string;
  user_id: {
    $oid: string;
  };
  time_read?: {
    $date: number;
  };
}